move http HeaderString to common code tree for common usage
