+++ envoy/common/BUILD
@@ -106,3 +106,18 @@ envoy_cc_library(
     name = "hashable_interface",
     hdrs = ["hashable.h"],
 )
+
+envoy_cc_library(
+    name = "union_string",
+    hdrs = [
+        "union_string.h",
+    ],
+    external_deps = [
+        "abseil_inlined_vector",
+        "abseil_variant",
+    ],
+    deps = [
+        "//source/common/common:assert_lib",
+        "//source/common/common:utility_lib",
+    ],
+)
